Extract stock save, load and clear operations from Form to a dedicated object
